Change mf_fps to double
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@25996 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ libmpdemux/mf.c
@@ -25,7 +25,7 @@
 
 int    mf_w = 0; //352; // let codecs to detect it
 int    mf_h = 0; //288;
-float  mf_fps = 25.0;
+double mf_fps = 25.0;
 char * mf_type = NULL; //"jpg";
 
 mf_t* open_mf(char * filename){
